fix(host): make size_t prints flag-independent (cast to %llu)
The prior %z fix relied on -D__USE_MINGW_ANSI_STDIO=1 selecting
mingw's C99 stdio. On the user's mingw that macro does NOT switch the
implementation (printf still resolves to msvcrt), so the %zu warnings
persisted — and with -Werror now on, they became hard build errors
([-Werror=format=]). The flag was a no-op on that toolchain.

Fix at the source instead, so it's correct on ANY C runtime with no
flag dependency: replace %zu with %llu and cast the size_t arguments
to (unsigned long long) at the five host.c sites (apply_inicfg x2,
main CLI range checks x2, vm_system_init failure message). %zu was
valid C99 but depended on a runtime that honors it; the cast does
not depend on anything.

Verified the casts alone fix it: native mingw build is clean under
-Werror EVEN WITHOUT __USE_MINGW_ANSI_STDIO — so this works on the
user's mingw regardless of how it resolves that macro. Kept the flag
in the build scripts anyway (harmless, slightly better stdio
conformance where honored); the correctness now comes from the
casts, not the flag.

Swept the tree: no remaining %z specifiers in our source (FatFs is
vendored and compiled clean).

This is the warnings-policy "fix the code" path (option 1), not a
suppression — the casts are a genuine portability improvement for a
project targeting multiple runtimes.

Verified: native build clean under -Werror (with and without the
flag); Linux examples clean under -Werror; 15/15 vm, 4/4 real_elf,
6/6 stdio, 16/16 fs.

+++ examples/05_shell/host.c
@@ -259,7 +259,8 @@ static bool apply_inicfg(const IniCfg *cfg, HostConfig *hc) {
     if (inicfg_get_int(cfg, "memory", "local_kb", &lv)) {
         if (lv < 32 || lv > (long)(LOCAL_BYTES / 1024)) {
             fprintf(stderr, "vm.cfg: [memory] local_kb=%ld out of range "
-                    "(32..%zu)\n", lv, (size_t)(LOCAL_BYTES / 1024));
+                    "(32..%llu)\n", lv,
+                    (unsigned long long)(LOCAL_BYTES / 1024));
             return false;
         }
         hc->local_bytes = (size_t)lv * 1024;
@@ -267,7 +268,8 @@ static bool apply_inicfg(const IniCfg *cfg, HostConfig *hc) {
     if (inicfg_get_int(cfg, "memory", "shared_kb", &lv)) {
         if (lv < 8 || lv > (long)(SHARED_BYTES / 1024)) {
             fprintf(stderr, "vm.cfg: [memory] shared_kb=%ld out of range "
-                    "(8..%zu)\n", lv, (size_t)(SHARED_BYTES / 1024));
+                    "(8..%llu)\n", lv,
+                    (unsigned long long)(SHARED_BYTES / 1024));
             return false;
         }
         hc->shared_bytes = (size_t)lv * 1024;
@@ -1391,16 +1393,16 @@ int main(int argc, char **argv) {
      * silently clamped or wrapping). */
     if (cli_local_kb >= 0) {
         if (cli_local_kb < 32 || cli_local_kb > (long)(LOCAL_BYTES / 1024)) {
-            fprintf(stderr, "host: --local-kb=%ld out of range (32..%zu)\n",
-                    cli_local_kb, (size_t)(LOCAL_BYTES / 1024));
+            fprintf(stderr, "host: --local-kb=%ld out of range (32..%llu)\n",
+                    cli_local_kb, (unsigned long long)(LOCAL_BYTES / 1024));
             return 1;
         }
         hc.local_bytes = (size_t)cli_local_kb * 1024;
     }
     if (cli_shared_kb >= 0) {
         if (cli_shared_kb < 8 || cli_shared_kb > (long)(SHARED_BYTES / 1024)) {
-            fprintf(stderr, "host: --shared-kb=%ld out of range (8..%zu)\n",
-                    cli_shared_kb, (size_t)(SHARED_BYTES / 1024));
+            fprintf(stderr, "host: --shared-kb=%ld out of range (8..%llu)\n",
+                    cli_shared_kb, (unsigned long long)(SHARED_BYTES / 1024));
             return 1;
         }
         hc.shared_bytes = (size_t)cli_shared_kb * 1024;
@@ -1556,10 +1558,11 @@ int main(int argc, char **argv) {
     };
     if (!vm_system_init(&sys, &cfg)) {
         fprintf(stderr, "host: vm_system_init failed "
-                "(local=%zu shared=%zu max_vms=%u spawn_data_kb=%u — "
+                "(local=%llu shared=%llu max_vms=%u spawn_data_kb=%u — "
                 "try smaller values in vm.cfg or via --max-vms / "
                 "--spawn-data-kb)\n",
-                hc.local_bytes, hc.shared_bytes,
+                (unsigned long long)hc.local_bytes,
+                (unsigned long long)hc.shared_bytes,
                 (unsigned)hc.max_vms, (unsigned)hc.spawn_data_kb);
         return 1;
     }
